pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.124.0",
    "openai>=2.9.0",
    "openai-agents>=0.6.2",
    "python-dotenv>=1.2.1",
    "python-env>=1.0.0",
    "qdrant-client>=1.16.1",
    "uvicorn>=0.38.0",
]
